zoom_testbed/requirements.txt
opencv
numpy>=1.22.4
pandas
pylibdmtx
scikit-image
pillow
ffmpeg
